22-12-22/CRUD_TEMPLATE3/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-appconf==1.0.5
django-imagekit==4.1.0
pilkit==2.0
Pillow==9.3.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
